Add files via upload
This is the most basic stepper motor driver.  As it sits, it will rotate a stepper motor connected to a DRV8825 driver and an Arduino board and spin one revolution at all of the available micro steps.  Of course it will work with a Mega and variations of an RP2040 board, such as the Pi Pico.

I really recommend inserting a 100 mF capacitor across the VMOT and GND, as it will really quieten and smooth out the pulses.